+++ docs/changelog.rst
@@ -227,6 +227,11 @@ Detailed list of changes
 
 - A new action :ac:`copy_selection_or_last_command_output` (:pull:`9512`)
 
+- Wayland: Add support for the background blur extension (:iss:`9534`)
+
+- macOS: A new option :opt:`macos_dock_badge_on_bell` to show a badge on the
+  kitty dock icon when a bell occurs (:pull:`9529`)
+
 
 0.45.0 [2025-12-24]
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~